:white_check_mark: 更新测试代码

+++ src/main/java/com/zmy/controller/ProductController.java
@@ -4,25 +4,33 @@ import com.zmy.dto.Message;
 import com.zmy.entity.Product;
 import com.zmy.service.ProductService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpRequest;
 import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.*;
+
+import javax.annotation.Resource;
+import javax.servlet.http.HttpServletRequest;
 
 /**
  * 控制层
  * controller层 = web层
  * 主要负责具体业务模块流程的控制，需要调用service逻辑设计层的接口来控制业务流程
  * 因为service中的方法是我们使用到的，controller通过接收前端传过来的参数进行业务操作，在返回一个指定的路径或者数据表
+ * @author zmy
  */
 //@Controller用于标注控制层组件
-@Controller
 //处理请求地址映射，可用于类或方法上。用于类上，表示类中的所有响应请求的方法都是以该地址作为父路径
+
 @RequestMapping("/product")
+@Controller
 public class ProductController {
     //属性加上@Autowired后不需要getter()和setter()方法，Spring也会自动注入。
     //@Autowired注释使得接口可以被容器注入
-    @Autowired
+
+    @Resource
+    private HttpServletRequest request;
+
+    @Resource
     private ProductService productService;
 
     /**
@@ -31,12 +39,13 @@ public class ProductController {
      * @param proId
      * @return
      */
-    @RequestMapping(value = "inquireProductById")
     //@ResponseBody注解的作用是将controller的方法返回的对象通过适当的转换器转换为指定的格式之后
     //写入到response对象的body区，通常用来返回JSON数据或者是XML数据
-    @ResponseBody
     //@RequestParam用于将指定的请求参数赋值给方法中的形参
-    public Message getProductById(@RequestParam(value = "proId") Integer proId) {
+
+    @RequestMapping(value = "inquireProductById", method = RequestMethod.GET)
+    @ResponseBody
+    public Message inquireProductById(@RequestParam(value = "proId") Integer proId) {
         return Message.success().addObject("product", productService.inquireProductById(proId));
     }
 
@@ -55,9 +64,11 @@ public class ProductController {
                               @RequestParam(value = "proPrice") Double proPrice,
                               @RequestParam(value = "proInventory") Integer proInventory,
                               @RequestParam(value = "merId") Integer merId) {
-        productService.addProduct(new Product(proName, proPrice, proInventory, merId));
-        return Message.success("添加成功");
-
+        Integer res = productService.addProduct(new Product(proName, proPrice, proInventory, merId));
+        if (res > 0){
+            return Message.success("添加成功");
+        }
+        return Message.error("添加失败");
     }
 
     /**
@@ -66,7 +77,6 @@ public class ProductController {
      * @param proName
      * @param proPrice
      * @param proInventory
-     * @param merId
      * @return
      */
     @RequestMapping(value = "modifyProInfo")
@@ -74,17 +84,29 @@ public class ProductController {
     public Message inquireProInfo(@RequestParam(value = "proId") Integer proId,
                                   @RequestParam(value = "proName") String proName,
                                   @RequestParam(value = "proPrice") Double proPrice,
-                                  @RequestParam(value = "proInventory") Integer proInventory,
-                                  @RequestParam(value = "merId") Integer merId) {
-        productService.modifyProInfo(new Product(proName, proPrice, proInventory, merId));
-        return Message.success("修改成功");
+                                  @RequestParam(value = "proInventory") Integer proInventory) {
+
+        Integer merId = (Integer) request.getSession().getAttribute("merId");
+        Integer res = productService.modifyProInfo(new Product(proId, proName, proPrice, proInventory, merId));
+        if (res > 0){
+            return Message.success("修改成功");
+        }
+        return Message.error("修改失败");
 
     }
 
-    @RequestMapping(value = "deleteProduct")
+    @RequestMapping(value = "deleteProduct",method = RequestMethod.DELETE)
     @ResponseBody
     public Message deleteProduct(@RequestParam(value = "proId") Integer proId) {
-        productService.deleteProduct(proId);
-        return Message.success("删除成功");
+        Integer res = 0;
+        Integer merId = productService.inquireProductById(proId).getMerId();
+        final String key = "merId";
+        if (merId == request.getSession().getAttribute(key)){
+            res = productService.deleteProduct(proId);
+        }
+        if (res > 0){
+            return Message.success("删除成功");
+        }
+        return Message.error("删除失败");
     }
 }
